Adding the final report and the presentation
The Final report consists of in-depth analysis about our experimentation, architecture, tools and other things we built and did our research on. 
Presentation is the final class project presentation, which include high level understanding of the work done, things we experimented, final result, economic value and the future work that needs to be done.